Remove duplicate `AttachmentPreviewProgressStyle`
